move prepare_settings_under_seed to cezo_fl.util.prepare_settings
